Replace pizza tooltip with accordion
